requirements.txt
pillow
flask